Fix race in mux.emitMetrics (fixes #308)
